v1.15.2 - 20170310 Fixed AI being broken with vanilla starting tile due to a change in v1.15.1 Can no longer start an AI game with City of Carcassonne tileset Can no longer start a game without a starting tileset
v1.15.1 - 20170228
Fixed Barn feature markers, which were not working. Also made the marker appear on the corner instead of the standard position markers, which would have made little sense to the player.
Fixed feature markers showing invalid locations for special figures when using a tile with a magic portal (portals are only supposed to work for followers).